refactor: address CR review iter-1 — 判定を純関数に切り出す (#3023)
`warnAboutUnwiredRoots` が 23 行あった。指摘は正しい。

`missingRootCapabilities(backend)` として `types.ts` に export した
（8 / 17 行）。CLAUDE.md の「純ロジックは export したヘルパへ分離して
テスト可能にする」にも合致する ——— **サーバを組み立てずに規則そのものを
駆動できる**ようになったので、テストを 5 件足した。

うち 1 件は `rootScopedGenerationState: false` の扱い。ホストが「scope できない」
と明示的に言った場合で、何も言わないのと同じ答えでなければならない。
これは自分で足したもので、CodeRabbit の指摘には無い。

テスト 339 件（+5）。
