+++ packages/plugins/mulmoscript-plugin/src/server/ops.ts
@@ -48,6 +48,7 @@ import { normalizeStoryPath, storyRefWithin, storiesRelativePath } from "../core
 import { errorMessage } from "@mulmoclaude/common";
 import { resolveWithinRoot } from "@mulmoclaude/core/files";
 import { fileToDataUri, stripDataUri } from "./support";
+import { missingRootCapabilities } from "./types";
 import { enableGraphAIErrorCapture, setMulmoErrorCaptureLogger, withMulmoErrorCapture } from "./mulmoErrorCapture";
 import type {
   GenerateOpArgsWith,
@@ -240,22 +241,7 @@ export function createMulmoScriptServerOps(backend: MulmoScriptServerBackend) {
    */
   function warnAboutUnwiredRoots(): void {
     if (rootDirs.size <= 1) return;
-    const missing: string[] = [];
-    // The pair identity is complete INSIDE this package and stops at the host
-    // boundary: a host's per-session store keys pending work on
-    // `(kind, filePath, key)` — `generationKey` in `@mulmobridge/protocol` —
-    // so two roots generating the same beat in one session collapse to one
-    // entry (Codex P1 on #3015). A host that keeps no such store declares it.
-    if (backend.rootScopedGenerationState !== true) {
-      missing.push(
-        "GENERATION is refused until this host declares `rootScopedGenerationState` (its pending-generation state must carry the root, or it must keep none)",
-      );
-    }
-    // The save/update executors run against one FileOps; without a per-root
-    // one they would rewrite the DEFAULT root's identically-named script.
-    if (backend.artifactsFor === undefined) {
-      missing.push("save/update land in the DEFAULT root until this host passes `artifactsFor`");
-    }
+    const missing = missingRootCapabilities(backend);
     if (missing.length === 0) return;
     log.warn(`extra stories roots registered — reads and uploads work, but ${missing.join(", and ")} (#3019)`, {
       roots: [...rootDirs.keys()].filter((id) => id !== DEFAULT_ROOT),
